requirements.txt
requests>=2.25
python-dotenv==0.17